requirements.txt
asgiref==3.2.7
cffi==1.14.0
cryptography==2.8
Django==3.0.4
django-cors-middleware==1.5.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
psycopg2-binary==2.8.4
pycparser==2.20
PyJWT==1.7.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
